requirements.txt
flask
gunicorn
requests
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
responses>=0.23.0
